+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix: Fix the bug that Toastx.show\*(\*, int) method may not get multilingual information
 * fix: Fix the bug that there may be null elements in the array returned by Storagex.getAppExternal\*Dirs() method
 * new: Storagex.clean\*Dirs() method now returns boolean
+* change: LifecycleBroadcastReceiver can now register multiple times
 
 ## v1.0.0-alpha09
 * remove: Remove dependency on javax